requirements.txt
cryptography==2.9.2
boto==2.49.0
boto3==1.16.37
botocore==1.19.37
dash==1.16.3
dash-bootstrap-components==0.10.7
dash-core-components==1.12.1
dash-html-components==1.1.1
dash-renderer==1.8.2
dash-table==4.10.1
click==7.1.2
Flask==1.1.2
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
unicodecsv==0.14.1
Werkzeug==1.0.1
python-dateutil==2.8.1
Flask-SQLAlchemy==2.4.4
mysql-connector-python==8.0.21
msgpack==1.0.0
appnope==0.1.0
backcall==0.2.0
Brotli==1.0.9
certifi==2020.6.20
PyJWT==1.7.1
coverage==5.3
decorator==4.4.2
Flask-Compress==1.7.0
future==0.18.2
ipython==7.18.1
ipython-genutils==0.2.0
jedi==0.17.2
numpy==1.19.3
pandas==1.1.3
parso==0.7.1
pexpect==4.8.0
pickleshare==0.7.5
plotly==4.12.0
prompt-toolkit==3.0.8
ptyprocess==0.6.0
Pygments==2.7.2
PyMySQL==0.10.1
pytz==2020.1
retrying==1.3.3
six==1.15.0
SQLAlchemy==1.3.20
traitlets==5.0.5
wcwidth==0.2.5